+++ bears_ecology/src/request.rs
@@ -1,4 +1,4 @@
-use crate::{App, History, Mode, Options, Queue, bea_data, init};
+use crate::{App, History, Mode, Options, Overwrite, Queue, Scope, bea_data, init};
 use bears_species::{
     BeaErr, BeaResponse, Data, Dataset, DatasetMissing, FixedAssets, GdpByIndustry, IoError, Ita,
     Method, Mne, NiUnderlyingDetail, Nipa, ParameterName, ReqwestError, Results, SerdeJson,
@@ -62,7 +62,7 @@ fn init_method(method: Method) -> Result<App, BeaErr> {
     let mut app = init()?;
     tracing::info!("App initialized.");
     let mut options = Options::default();
-    options.with_method(method);
+    let _ = options.with_method(method);
     app.with_options(options);
     tracing::info!("App configured for {method}.");
     Ok(app)
@@ -157,7 +157,7 @@ pub async fn get_datasets() -> Result<(), BeaErr> {
 pub async fn initial_download(dataset: Dataset) -> Result<(), BeaErr> {
     let queue = init_queue(dataset)?;
     tracing::info!("Queue length: {}", queue.len());
-    queue.download(false).await?;
+    queue.download(Overwrite::No).await?;
     Ok(())
 }
 
@@ -172,7 +172,10 @@ pub async fn download_with_history(
     // get the download history for the size hints
     let history = History::try_from((dataset, Mode::Load))?;
     history.summary();
-    history.iter().download(&queue, false, style).await?;
+    history
+        .iter()
+        .download(&queue, Overwrite::No, style)
+        .await?;
     Ok(())
 }
 
@@ -191,7 +194,7 @@ pub fn initial_load_par(
     let downloads = History::try_from((dataset, Mode::Download))?;
     // only download successes in history
     // strict = true set to include no others in queue.
-    queue.successes(&downloads, true)?;
+    queue.successes(&downloads, Scope::History)?;
     tracing::info!("Files downloaded: {}", queue.len());
 
     if let Some(loads) = load_history {
@@ -220,7 +223,7 @@ pub async fn initial_load(
     let downloads = History::try_from((dataset, Mode::Download))?;
     // only download successes in history
     // strict = true set to include no others in queue.
-    queue.successes(&downloads, true)?;
+    queue.successes(&downloads, Scope::History)?;
     tracing::info!("Files downloaded: {}", queue.len());
 
     if let Some(loads) = load_history {
@@ -244,7 +247,7 @@ pub async fn retry_load(dataset: Dataset) -> Result<Vec<Data>, BeaErr> {
     // The load history contains errors, try them again.
     let history = History::try_from((dataset, Mode::Load))?;
     // strict is true means only download errors included in the event history
-    queue.errors(&history, true)?;
+    queue.errors(&history, Scope::History)?;
     tracing::info!("Files to retry: {}", queue.len());
 
     let data = queue.load().await?;
@@ -283,7 +286,7 @@ async fn parameter(dataset: Dataset, app: &mut App) -> Result<(), BeaErr> {
             let method = "get".to_string();
             let body = app.params().into_iter().collect::<Vec<(String, String)>>();
             let mut error = ReqwestError::new(url, method, source, line!(), file!().to_string());
-            error.with_body(body);
+            let _ = error.with_body(body);
             Err(error.into())
         }
     }
@@ -313,8 +316,8 @@ async fn parameter_value(
     name: ParameterName,
 ) -> Result<(), BeaErr> {
     let mut opts = app.options().clone();
-    opts.with_dataset(dataset);
-    opts.with_param_name(name);
+    let _ = opts.with_dataset(dataset);
+    let _ = opts.with_param_name(name);
     app.with_options(opts);
     let data = app.get().await?;
     match data.json::<serde_json::Value>().await {
@@ -339,7 +342,7 @@ async fn parameter_value(
             let method = "get".to_string();
             let body = app.params().into_iter().collect::<Vec<(String, String)>>();
             let mut error = ReqwestError::new(url, method, source, line!(), file!().to_string());
-            error.with_body(body);
+            let _ = error.with_body(body);
             Err(error.into())
         }
     }
@@ -380,8 +383,8 @@ pub async fn parameter_values() -> Result<(), BeaErr> {
 #[tracing::instrument(skip_all)]
 async fn value(dataset: Dataset, app: &mut App, name: ParameterName) -> Result<(), BeaErr> {
     let mut options = app.options().clone();
-    options.with_dataset(dataset);
-    options.with_target(name);
+    let _ = options.with_dataset(dataset);
+    let _ = options.with_target(name);
     app.with_options(options.clone());
     let data = app.get().await?;
     tracing::info!("{data:#?}");
@@ -420,7 +423,7 @@ async fn value(dataset: Dataset, app: &mut App, name: ParameterName) -> Result<(
             let method = "get".to_string();
             let body = app.params().into_iter().collect::<Vec<(String, String)>>();
             let mut error = ReqwestError::new(url, method, source, line!(), file!().to_string());
-            error.with_body(body);
+            let _ = error.with_body(body);
             return Err(error.into());
         }
     }
@@ -498,8 +501,8 @@ async fn value_gdp(dataset: Dataset, app: &mut App, name: ParameterName) -> Resu
     };
     // Add dataset and target parameter to options
     let mut options = app.options().clone();
-    options.with_dataset(dataset);
-    options.with_target(name);
+    let _ = options.with_dataset(dataset);
+    let _ = options.with_target(name);
     // navigate to parameter_values directory
     let path = bea_data.join("parameter_values");
     // create the folder if it does not exist
@@ -521,7 +524,7 @@ async fn value_gdp(dataset: Dataset, app: &mut App, name: ParameterName) -> Resu
         ParameterName::Industry => {
             for id in table_id {
                 // add table id to options
-                options.with_table_id(*id.value());
+                let _ = options.with_table_id(*id.value());
                 // update app with modified options
                 app.with_options(options.clone());
                 // fire off the get request using the configured app
@@ -550,7 +553,7 @@ async fn value_gdp(dataset: Dataset, app: &mut App, name: ParameterName) -> Resu
                         let body = app.params().into_iter().collect::<Vec<(String, String)>>();
                         let mut error =
                             ReqwestError::new(url, method, source, line!(), file!().to_string());
-                        error.with_body(body);
+                        let _ = error.with_body(body);
                         return Err(error.into());
                     }
                 }
@@ -559,7 +562,7 @@ async fn value_gdp(dataset: Dataset, app: &mut App, name: ParameterName) -> Resu
         // TODO: Test this branch
         ParameterName::Year => {
             for id in table_id {
-                options.with_table_id(*id.value());
+                let _ = options.with_table_id(*id.value());
                 app.with_options(options.clone());
                 let data = app.get().await?;
                 tracing::info!("{data:#?}");
@@ -582,7 +585,7 @@ async fn value_gdp(dataset: Dataset, app: &mut App, name: ParameterName) -> Resu
                         let body = app.params().into_iter().collect::<Vec<(String, String)>>();
                         let mut error =
                             ReqwestError::new(url, method, source, line!(), file!().to_string());
-                        error.with_body(body);
+                        let _ = error.with_body(body);
                         return Err(error.into());
                     }
                 }